requirements.txt
as
cachetools
certifi
chardet
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
httplib2
idna
numpy
oauth2client
packaging
pandas
pd
protobuf
psycopg2-binary
pyasn1
pyasn1-modules
PyDrive
pyparsing
python-dateutil
pytz
PyYAML
requests
rsa
six
uritemplate
urllib3


